refactor(hotel): ScheduleGridView をコンポーネント分割して整理
855行に膨らんだ単一ファイルを、振る舞いを変えずに schedule-grid/ 配下へ分割。
ScheduleGridView は派生データの算出と編集状態の保持に徹するオーケストレーターに。

- styles.ts: 巨大な StyledWrap(CSS) を分離
- ScheduleDataCell: 1セルの純粋表示
- StaffScheduleRow: スタッフ1行（ヘッダ＋セル＋休合計＋希望行）
- SummaryRow: 集計行1行（充足率バー・休合計）
- EditMenus: セル編集／必要人数編集メニュー
- summaryModel / wishSummary / constants / types: 派生モデルと共有定義
